+++ packages/terra-open-source-scripts/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Changed
   * Update to use `prompts` instead of `inquirer`. It has a smaller memory footprint
+  * Updated to use a simpler import of the terra-cli logger
 
 ## 1.1.0 - (November 19, 2020)