+++ packages/webpack-config-terra/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Updated dependencies.
+
 ## 1.0.0-alpha.1 - (October 20, 2020)
 
 * Changed